Change Todo struct to a class
This change was implemented so objects will be able to perform todo
related tasks without the functions and variables being seperated. In
addition, code was moved from main to functions that require it in
order to work as intended and variables have default values rather than
relying on the user to set them.